split version subcommands plus docs
