+++ docs/cmd/version/check.md
@@ -1,0 +1,21 @@
+# check
+
+## Description
+
+The `version check` command checks if a newer version of containerlab is available.
+
+## Usage
+
+```
+containerlab version check
+```
+
+When executed, the containerlab binary will reach out to the github API and check if a newer version is available and display the result.
+
+Example result output:
+
+```
+🎉 A newer containerlab version (0.62.2) is available!
+Release notes: https://containerlab.dev/rn/0.62/#0622
+Run 'sudo clab version upgrade' or see https://containerlab.dev/install/ for installation options.
+```
