+++ docs/cmd/version/index.md
@@ -1,0 +1,19 @@
+# version
+
+## Description
+
+The `version` command displays the version of containerlab installed.
+
+## Usage
+
+```
+containerlab version
+```
+
+The version command contains:
+
+* version in the semantic versioning format
+* commit hash of the project repository
+* build date
+* a link to the github repo
+* a link to the release notes of the current version
