+++ mkdocs.yml
@@ -104,6 +104,9 @@ nav:
           - netem:
               - set: cmd/tools/netem/set.md
               - show: cmd/tools/netem/show.md
+      - version:
+          - cmd/version/index.md
+          - check: cmd/version/check.md
       - completions: cmd/completion.md
   - Lab examples:
       - About: lab-examples/lab-examples.md
